+++ tests/availableCores.R
@@ -45,11 +45,28 @@ stopifnot(inherits(res, "try-error"))
 ncores0 <- 42L
 
 message("*** LSF ...")
+message(" - LSB_DJOB_NUMPROC")
 Sys.setenv(LSB_DJOB_NUMPROC = as.character(ncores0))
 ncores <- availableCores(methods = "LSF")
 print(ncores)
 stopifnot(ncores == ncores0)
 message("*** LSF ... done")
+
+message("*** PJM (Fujitsu Technical Computing Suite) ...")
+message(" - PJM_VNODE_CORE")
+Sys.setenv(PJM_VNODE_CORE = as.character(ncores0))
+ncores <- availableCores(methods = "PJM")
+print(ncores)
+stopifnot(ncores == ncores0)
+Sys.unsetenv("PJM_VNODE_CORE")
+
+message(" - PJM_PROC_BY_NODE")
+Sys.setenv(PJM_PROC_BY_NODE = as.character(ncores0))
+ncores <- availableCores(methods = "PJM")
+print(ncores)
+stopifnot(ncores == ncores0)
+Sys.unsetenv("PJM_PROC_BY_NODE")
+message("*** PJM (Fujitsu Technical Computing Suite) ... done")
 
 
 message("*** Internal detectCores() ...")