+++ NEWS.md
@@ -2,9 +2,10 @@
 
 ## New Features
 
- * `availableCores()` gained support for the
-   Fujitsu Technical Computing Suite.  Specifically, it acknowledges
-   environment variables `PJM_VNODE_CORE` and `PJM_PROC_BY_NODE`.
+ * `availableCores()` and `availableWorkers()` gained support for the
+   'Fujitsu Technical Computing Suite' job scheduler.  Specifically, they
+   acknowledges environment variables `PJM_VNODE_CORE`, `PJM_PROC_BY_NODE`,
+   and `PJM_O_NODEINF`.
 
 ## Bug Fixes